arm positioning and safety precautions
today we read the encoders of our arm to use the encoder values and its height off the ground in inches to put in a linear equation in which we found the slope and offset, which allowed us to show the inch value of our arms positions instead of encoder counts. we also with this put in safeties to make sure our arm doesn't exceed its maximum height and slip the belt.